Minor cleanup
- Cleanup ContentMannager: Put template method in header
- Sse std::type_index instead of plain string to store type value in ServiceProvidder
- Add bat class

+++ src/GameCuaTao/Direct2DGame/Game.cpp
@@ -84,7 +84,7 @@ void Game::Run()
 	{
 		Tick();
 		auto deltaTime = (int)gameTime.ElapsedGameTime.Milliseconds();
-		auto tickPerFrame = graphics->GetTickPerFrame();
+		auto tickPerFrame = (int)graphics->GetTickPerFrame();
 
 		if (deltaTime >= tickPerFrame)
 		{
